+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import './home_page.dart';
-import './navigation_bar.dart';
+import 'pages/home_page.dart';
 import './routes.dart';
 
 void main() {
@@ -10,10 +9,6 @@ void main() {
 class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-        home: Scaffold(
-      body: HomePage(),
-      bottomNavigationBar: NavigationBar(),
-    ), routes: routes);
+    return MaterialApp(home: HomePage(), routes: routes);
   }
 }
