+++ lib/pages/booking_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../navigation_bar.dart';
 
 class BookingPage extends StatelessWidget {
   @override
@@ -7,6 +8,7 @@ class BookingPage extends StatelessWidget {
       body: Row(
         children: [Text('Booking page')],
       ),
+      bottomNavigationBar: NavigationBar(),
     );
   }
 }
