+++ lib/pages/my_message_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../navigation_bar.dart';
 
 class MyMessagePage extends StatelessWidget {
   @override
@@ -7,6 +8,7 @@ class MyMessagePage extends StatelessWidget {
       body: Row(
         children: [Text('my message page')],
       ),
+      bottomNavigationBar: NavigationBar(),
     );
   }
 }
