+++ lib/routes.dart
@@ -1,8 +1,8 @@
-import './booking_page.dart';
-import './search_page.dart';
-import './create_trip_page.dart';
-import './my_message_page.dart';
-import './my_account_page.dart';
+import 'pages/booking_page.dart';
+import 'pages/search_page.dart';
+import 'pages/create_trip_page.dart';
+import 'pages/my_message_page.dart';
+import 'pages/my_account_page.dart';
 import 'package:flutter/material.dart';
 
 Map<String, Widget Function(BuildContext)> routes = {
